Add: many tools and functions
Various more C functions have been added and now printf has a C function
as well (had a "special" idea to have that working). I've also
refactored almost the entire project to get it to a more presentable
(and understandable) state.

Now it's possible to use all functions normally and not worry too much
about Assembly (except on new functions for the EXE which must be
written in Assembly really.

I'm now opening an INI and reading a property value from there.

The realMain() now returns a bool also to say everything went fine or
not - will need a small change on the loader, which will come in the
next commit right away.