Add: Loader freeing code block and more
Now the Loader aside from loading, also has code to unload the allocated
code block. The address is stored in the end of the Code section of the
game EXE. When main_() ends, there's a call to exit(). I'm redirecting
that call to the part of the Loader that unloads the code block.

Aside from that, the Loader now checks the BIN file length when looking
for realMain(), so that if that function is not found, it doesn't crash
and so it reports an error and is able to keep loading the game without
any patches (since the Patcher didn't get to start).