Add: Loader EXE patches to the Installer
I've also removed the permanent patches that were on it, except the
dialog_out_ function, which is still useful. Will be taken care of some
time later - will give much work to put the rest there too...

Aside from that, the naming is corrected here and the credits too.

It's just missing the Loader and Patcher bytes on it. Everything else I
think it's ready.