Update: _WinMain_hook hook's location
I've put NOPs on the beginning of the Loader code on the Loader exactly
to be possible to hook something there, right before the block of code
is freed by free(). So this is now corrected. If I'd leave it as it was,
it would be disabling the free() call.