Add: more DllMain2() patches
Also, I've put all INI intever values with sscanf() on them so that it's
possible to put a value of "1 " and that being the same as "1" or "1  ".
Before, I was just checking if the value was equal to "1" and that's not
a good idea, so string to int conversion applied.

Note: I think the PartyControl patches are not working. But I'll leave
the code active to see if other patches are needed or something. I
already reviewed all the code and it seems exactly equal to Crafty's.