Update: Loader
Now it free()s and close()s what it should but it did not previously,
not leaving any memory leaks, supposedly.

I've also fixed the problem with it not showing an error string
correctly.

Aside from that, a small issue with a push instead of a pop in the
unload part.

BIN file name was also changed to f1dpatch.bin, and the initial and end
strings updated.