Update: Installer is now ready
The Installer is now ready with the Loader bytes on it.

The Data section will not be set as Executable because that will crash
DOS/32A even though I believe it should not. It doesn't crash DOS/4GW,
but then it would make the game not playable on DOS/32A by a thing which
doesn't even matter because the LE loader seems to ignore these flags,
so I disabled that patch for the game to be playable on DOS/32A.

The Installer will also not create any files. All will go in a ZIP. It
will only install the Loader into the game EXE and nothing else.

I've also deleted the function patches because they were for me to see
how to get Fallout Fixt 7.0alpha an newer to work on MS-DOS. I still
think it has to do with something different between the MS-DOS and the
Windows EXEs, but I've got it working without needing to mess with the
EXE, so no more need for the function patches.