Update: Crafty's sFall1 base version is now 1.7.20
F1DP Patcher version increased to 1.1 because of this.

Also fixed bugs that happened because of my lack of attention on the ASM
code. So, some things that probably were not working were fixed.

Removed the freeNew macro because of attempting to write to a variable
without correcting its address. Not needed here anyway, I almost don't
allocate memory dinamically.