Correct custom compiler and C99 code update
It seems Watcom can understand C99 through -za99. So I added the flag
and updated the code - on the Loader only though. No way I'm updating
the Patcher's code just for this... I also removed the -ze flag for
compiler extensions (no usage for it anyway).

Also corrected the custom compiler YAML that configures Open Watcom 2 on
CLion. I remember I did some random testing on it and left it wrong, so
I downloaded the file again (and updated too, 3 months ago, it seems),
and just added the "code-insight-target-name: x86" to it (else CLion
doesn't provide all insight it can).